Add files via upload
This script takes out the 16S from barrnap and the blast output from it against the GTDB r220 16S database. Then, I extract the number of 16S copies in each barrnap output, and its blast output. Finally, I apply a 98.6% identity filter, a 1e-100 e-value, and a genus and species match to GTDBtk first, and then Sanger 16S (v3v4). This creates an output file of each blast output filtered, a finalized output file too that agglomerates all of the results, and finally, a fasta file ready to be used.